Float comparer fixes

+++ Programming/C#PartOne/HomeWorks/02.DataTypesAndVariables/ComparingFloats/FloatComparer.cs
@@ -6,10 +6,10 @@
     {
         static void Main(string[] args)
         {
-            var first = Math.Abs(float.Parse(Console.ReadLine()));
-            var second = Math.Abs(float.Parse(Console.ReadLine()));
+            var first = Math.Abs(double.Parse(Console.ReadLine()));
+            var second = Math.Abs(double.Parse(Console.ReadLine()));
             var diff = first - second;
-            float epsilon = 0.000001f;
+            double epsilon = 0.000001f;
             
             bool equal = Math.Abs(first) - Math.Abs(second) <= epsilon;
